Create radio module and example
This is the beginning of modeling our
code around our board design. Modules
are singletons that manage a specific hw context.
This makes importing easier and provides one layer of
abstraction between the interface an the hardware.

I provided a TX and RX example using the new system,
and creating a module based template is on my TODO.